requirements.txt
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
Pillow==11.0.0
django-filter==24.3
python-decouple==3.8
psycopg2-binary==2.9.10
drf-spectacular==0.28.0

# Testing
pytest==8.4.2
pytest-django==4.11.1
factory-boy==3.3.3
faker==37.12.0

